mcp: include scopes supported in the www-authenticate header (#1608)
**Description**

Add the configured scopes to the `WWW-Authenticate` headers. At
initialization time, which is when the first authentication will occur,
we don't have enough information to provide a fine-grained list of
scopes, so the best we can do is to default to the ones defined in the
protected resource metadata.

**Related Issues/PRs (if applicable)**

Fixes https://github.com/envoyproxy/ai-gateway/issues/1578

The addition of the header on 403 requests is implemented in
https://github.com/envoyproxy/ai-gateway/pull/1482, but this issue can
be closed as soon as this PR is merged, because we'll be compatible with
the latest spec.

**Special notes for reviewers (if applicable)**

cc @zhaohuabing can you take a look?

Signed-off-by: Ignasi Barrera <nacx@apache.org>