feat: initial tracing for Anthropic /messages API (#1611)
**Description**

This adds the initial implementation of distributed tracing for
Anthropic's /messages endpoint. To do that, this defines the concrete
API schema instead of generic map[string]any that had been used
previously.

**Related Issues/PRs (if applicable)**

* Redo of #1464 which was reverted in #1516
* Closes #1389 



___

Note: The implementation is passing all tests as well as the manual
integration test with Claude Code locally. The TODO is that i haven't
implemented all the input content block param type, so that and VCR
tests will be covered in the followup.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>