feat: cel based mcp authorization (#1684)
**Description**

This PR introudces a more flexible CEL-based MCP authorization, which
algins with the k8s networking agentic API proposal:
https://github.com/kubernetes-sigs/kube-agentic-networking/pull/17/changes
.

Unlike the Agentic API, our API is primarily intended for end-user
authorization. For this use case, a more type-safe and user-friendly
interface should probably be the primary option, since these policies
are more likely to be configured by an administrator after deployment
and tailored to individual users.

However, CEL is still a valuable addition as it provides a dynamic and
flexible way to configure authorization for more advanced use cases.

---------

Signed-off-by: Huabing Zhao <zhaohuabing@gmail.com>
Signed-off-by: Huabing(Robin) Zhao <zhaohuabing@gmail.com>
Co-authored-by: Ignasi Barrera <ignasi@tetrate.io>